Redid database service tests after database changes. Also changed long data type to int for consistency across all files.

+++ WorkoutApp/Models/CartItem.cs
@@ -9,10 +9,10 @@ namespace WorkoutApp.Models
 {
     public class CartItem
     {
-        public long Id { get; set; }
-        public long CartId { get; set; }
-        public long ProductId { get; set; }
-        public long Quantity { get; set; }
+        public int Id { get; set; }
+        public int CartId { get; set; }
+        public int ProductId { get; set; }
+        public int Quantity { get; set; }
 
         public IProduct GetProduct(ProductRepository repo)
         {
